Add Czech keyboard layouts
Two variants:

* Czech Standard (= qwertz)

* Czech qwerty

The accented letters layout corresponds to the UCW [1] layout shipped
with xkb, so we follow this precedent.

Like the DE layout, and unlike the US layout, the CZ layout has a
comma key and a narrower space bar. The added comma key also serves as
a visual balance to the added accents key, to keep the spacebar
centered.

The layouts have been tested manually on PinePhone.

[1] https://github.com/xkbcommon/libxkbcommon/blob/c60b77ea512bef92e481be38972b58dd71a34180/test/data/symbols/cz#L180