Add 'bg' layout
This is the Bulgarian (BDS) layout. I took the liberty to remove "э"
from the layout, as it is not part of the Bulgarian alphabet and it was
left there for historical reasons, also not to mess with the layouts for
physical keyboards. Removing it gives more space for the shift_l and
backspace keys.

I've also added the letter "small i with grave" to the special symbols, as
it is occasionally used in Bulgarian.